requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
load-dotenv==0.1.0
nominatim==0.1
psycopg2==2.9.10
python-dotenv==1.0.1
python-telegram-bot==21.6
requests==2.32.3
sniffio==1.3.1
urllib3==2.2.3
